Python Basics
Python Basics: Conditional Statements, Loops, and Data Structures

This repository is a beginner-friendly guide to understanding Python fundamentals. It covers essential concepts like conditional statements (if, elif, else), loops (for, while), and key data structures (tuples, strings, dictionaries, lists). The importance of indentation in Python is also highlighted, as it plays a critical role in code execution. This repo is perfect for anyone looking to build a solid foundation in Python programming.